Add files via upload
This GitHub folder contains the deliverables for my pizza sales data analytics project. It includes a comprehensive report detailing the analysis process and insights gained, an interactive dashboard created in Excel for visualizing sales data, an SQL file with the database queries used for data analysis, and a video presentation summarizing key findings and results. Explore the contents to gain insights into sales trends, top-selling products, and optimization opportunities for the pizza restaurant.